Fixed Update Presence Asset return type
Discord apparently sends the ID of the image sent back. This is totes
annoying and so I accounted for it with the new nullable unsigned long
values within the Assets field. When a Update Presence is now received,
it will instead of replacing the previous one, will merge them together.
The resulting presence from the merge will be exactly the received
presence, but with the keys of the last sent/received presence.