Fixed Unity Ghosting
Discord thinks the Unity.exe process is your game process (since its
grabbing it for its ID as the game is being executed within the
Unity.exe). So when you stop playing you game and if you just abort the
pipe connection, Discord will keep the rich presence enabled (because it
thinks you just temporarly lost connection to the pipe and plan to
connect again later).
Resulting in a "ghost". I believe it has been reported in the offical
API. It is literally a unity only bug, but its a good idea to send the
close signal on close anyways
Still deciding if I want to make it always work that way or make it a
configuration option, since it does delay the releasing of system
resources