Rewrote Byte Parsing
There was a potential for a arithmic overflow as the uint received from
Discord gets used to define an arrays length. Since this can be bigger
than a int it had the potential to cause a overflow. This **might** have
been the cause of the second part of issue #33 . This has been replaced
with a better system that utilises a temporary mem stream to read into.