Removed byte swapping
Discord apparently sends the data in system endian, not network endian.
This confused me and I had a check for BigEndian machines to reverse the
data received. Unfortuantly BigEndian machines are quiet rare nowadays
so I havn't ever come across this issue. Hopefully this will fix some of
the #33 bugs.