UNITY UPDATE: Fixed build crash.
Updated the unity project and the DiscordRPC_Unity.unitypackage. Please update!

When playign a build,, the DiscordRPC would break and had the potential to crash the game. This was first spotted in issue #38 . This has been fixed now. The cause of it was that the project was still in .NET 2.0 SUBSET which is not supported by Newtonsoft.Json or DiscordRPC. The fix was to add a dialog explicitly asking the user to convert the project.

I also fixed some silly file bugs with the native library. Removed the `.dll` at the end of all the `DLLImport` which means it will look for the assembly in the plugin folder instead of the CD. Also moved the UnityNamedPipe out into the project so it would be relative. In hindsight this is probably not required.

The new UnityNamedPipeClient now also uses Marshal to properly allocate the buffers for sending and receiving data. The Open also uses proper marshaling attributes for the string too now.

The FileLogger no longer crashes when JSON is feed into it. The Invite UI is back. Somehow it was moved but never comitted.

Refactored how the JSON parsing works a bit too, so it will be more consistent in the future.

Upgraded unity project to 2018, but it should be easily downgraded again. Was only doing this because its all I have on this machine atm.