Fixed Spelling & Inconsistencies. Added RP Builder.
Fixed some spelling mistakes on some functions, specifically Unubcribe
is now Unsubscribe. Fixed some inconsistencies and depreciated
functinos. Unsubscribe and Subscribe are obsolete in favour for
SetSubscription now. Added the Deinitialize function too to disable the
rich presence. Added the WithX().WithY() pattern to the Rich Presence
object for easier chaining. Obviously this is not the prefered way of
doing it, but some people like builders.