Refactored Constructors
The constructors have been consolidated into 2. They are now a simple
one that only takes appid and a advance one that takes all the
parameters, but most of them are optional. The RegisterUriScheme has
been seperated into its own function, removing both the boolean and the
steamID from the constructor.