Removed need to call Invoke()
Events will now automatically invoke themselves, removing the need for
the Invoke() function to be called regularly.

This functionality can be disabled and Invoke() can be called manually
in the constructor for situations where strict thread saftey is
paramount (eg: within the unity engine). The unity DiscordManager.cs
disables this and still uses Invoke as an example.