Prevented SetPresence Spam & Cloned Presence
If the SetPresence was called within a few milliseconds (less than
100ms) repeatedly, the pipe would not have enough time to process the
messages and will throw errors. This indicates an underlying flaw with
the mutexing of the pipe, but as a quick fix I have done the following:

Ignored presences that are identical to the current presence ( this can
be disabled using the `SkipIdenticalPresence` property )

Added a command queue size. This cannot be changed at the moment and
will only keep `512` _pending_ commands at a time.  Additional requests
will cause the oldest to be removed. In reality, this limit should never
be hit unless something horribly went wrong and someone is sending a new
presence every millisecond.

Another issue was that the stored current presence was not being cloned.
This resulted in situations where if you retained your copy of the Rich
Presence object, when a new one comes through discord that would have
been used instead.

Hopefully these changes will help with issue #71